pixly-backend/requirements.txt
asttokens==2.4.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
click==8.1.7
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar @ git+https://github.com/pallets-eco/flask-debugtoolbar@701219322023d67330395f4deb7bc796546c7181
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.7
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==2.3.8
WTForms==3.1.2
